Started Setting Refactoring
