+++ backend/src/repository/admin/mod.rs
@@ -8,14 +8,12 @@ pub fn init(cfg: &mut web::ServiceConfig) {
         .service(controller::list_repos_by_storage)
         .service(controller::list_repos)
         .service(controller::modify_security)
-        .service(controller::update_deployers_readers)
         .service(controller::modify_frontend_settings)
         .service(controller::modify_deploy)
         .service(controller::add_webhook)
         .service(controller::remove_webhook)
         .service(controller::update_active_status)
         .service(controller::update_policy)
-        .service(controller::clear_all)
         .service(controller::modify_badge_settings)
         .service(controller::update_description)
         .service(controller::delete_repository);
