+++ backend/src/repository/npm/models.rs
@@ -2,7 +2,10 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub  struct NPMSettings{
 
+}
 #[derive(Debug, Serialize, Deserialize)]
 pub struct LoginRequest {
     pub password: String,
